removed custom diff
