--- qbft/spectest/comparable/controller.go
@@ -1,12 +1,0 @@
-package comparable
-
-import (
-	"github.com/bloxapp/ssv-spec/qbft"
-	testingutilscomparable "github.com/bloxapp/ssv-spec/types/testingutils/comparable"
-)
-
-func CompareQBFTController(source, target *qbft.Controller) []testingutilscomparable.Difference {
-	ret := make([]testingutilscomparable.Difference, 0)
-
-	return ret
-}
