+++ ssv/spectest/tests/msg_processing_spectest.go
@@ -2,7 +2,6 @@ package tests
 
 import (
 	"encoding/hex"
-	testingutilscomparable "github.com/bloxapp/ssv-spec/types/testingutils/comparable"
 	"testing"
 
 	"github.com/bloxapp/ssv-spec/ssv"
@@ -60,7 +59,7 @@ func (test *MsgProcessingSpecTest) Run(t *testing.T) {
 	postRoot, err := test.Runner.GetRoot()
 	require.NoError(t, err)
 	if test.PostDutyRunnerStateRoot != hex.EncodeToString(postRoot[:]) {
-		testingutilscomparable.PrintDiff(test.Runner, RootRegister[test.PostDutyRunnerStateRoot])
+		testingutils.PrintDiff(test.Runner, RootRegister[test.PostDutyRunnerStateRoot])
 		require.Fail(t, "post runner state not equal")
 	}
 }
